RavenDB-16924 add failing test
